pyproject.toml
[tool.poetry]
name = "check_for_swf_in_file_list"
version = "0.1.0"
description = "check a file list to see if they are actually swf files or just fake pages"
authors = ["Mark Grandi <markgrandi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
attrs = "^20.3.0"
logging_tree = "^1.8.1"
arrow = "^0.17.0"

[tool.poetry.dev-dependencies]
wheel = "^0.36.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
